Create rsa.py
I created this as part of a group project for a cybersecurity course. The assignment involved implementing RSA and a factoring function. If I were to improve this code in the future, I would replace trial division with a more efficient sieve algorithm. I also listed references in the comments that were helpful in implementing in the CreateBlocks() and BlocksToMessage() functions.